+++ app/Http/Controllers/UserController.php
@@ -19,7 +19,8 @@ class UserController extends Controller
 {
     public function index(): JsonResponse
     {
-        $users = User::withTrashed()->get()->except(Auth::id());
+        $users = User::all()->except(Auth::id());
+        $deletedUsers = User::onlyTrashed()->get()->except(Auth::id());
 
         $users = $users->map(function ($user) {
             return [
@@ -35,14 +36,29 @@ class UserController extends Controller
             ];
         });
 
+        $deletedUsers = $deletedUsers->map(function ($user) {
+            return [
+                'id' => $user->id,
+                'name' => $user->name,
+                'email' => $user->email,
+                'roles' => $user->getRoleNames(),
+                'fechaNacimiento' => $user->fechaNacimiento->format('Y-m-d'),
+                'ci' => $user->ci,
+                'created_at' => $user->created_at->format('Y-m-d H:i:s'),
+                'updated_at' => $user->updated_at->format('Y-m-d H:i:s'),
+                'deleted_at' => $user->deleted_at == null ? null : $user->deleted_at->format('Y-m-d H:i:s'),
+            ];
+        });
+
         return response()->json([
             'data' => $users,
+            'deletedData' => $deletedUsers,
             'columns' => [
                 [ 'field' => 'id', 'header' => 'ID', 'type' => 'text'],
                 [ 'field' => 'name', 'header' => 'Nombre', 'type' => 'text'],
                 [ 'field' => 'ci', 'header' => 'Carnet de Identidad', 'type' => 'text'],
-                [ 'field' => 'email', 'header' => 'Email', 'type' => 'email'],
-                [ 'field' => 'role', 'header' => 'Rol', 'type' => 'chips'],
+                [ 'field' => 'email', 'header' => 'Email', 'type' => 'text'],
+                [ 'field' => 'roles', 'header' => 'Roles', 'type' => 'chips'],
                 [ 'field' => 'fechaNacimiento', 'header' => 'Fecha de Nacimiento', 'type' => 'text'],
                 [ 'field' => 'created_at', 'header' => 'Fecha de Creación', 'type' => 'text'],
                 [ 'field' => 'updated_at', 'header' => 'Ultima Actualización', 'type' => 'text'],
@@ -53,7 +69,7 @@ class UserController extends Controller
                 [ 'field' => 'email', 'header' => 'Email', 'type' => 'email' ],
                 [ 'field' => 'ci', 'header' => 'Carnet de Identidad', 'type' => 'text' ],
                 [ 'field' => 'fechaNacimiento', 'header' => 'Fecha de Nacimiento', 'type' => 'date' ],
-                [ 'field' => 'role', 'header' => 'Rol', 'type' => 'select', 'options' => Role::All()->pluck('name') ],
+                [ 'field' => 'roles', 'header' => 'Roles', 'type' => 'multiselect', 'options' => Role::All()->pluck('name') ],
                 [ 'field' => 'password', 'header' => 'Contraseña', 'type' => 'password' ],
                 [ 'field' => 'password_confirmation', 'header' => 'Confirmar Contraseña', 'type' => 'password' ],
             ],
@@ -62,7 +78,8 @@ class UserController extends Controller
                 [ 'field' => 'email', 'header' => 'Email', 'type' => 'email' ],
                 [ 'field' => 'ci', 'header' => 'Carnet de Identidad', 'type' => 'text' ],
                 [ 'field' => 'fechaNacimiento', 'header' => 'Fecha de Nacimiento', 'type' => 'date' ],
-                [ 'field' => 'role', 'header' => 'Rol', 'type' => 'select', 'options' => Role::All()->pluck('name') ],
+                [ 'field' => 'roles', 'header' => 'Roles', 'type' => 'multiselect', 'options' => Role::All()->pluck('name') ],
+                [ 'field' => 'password', 'header' => 'Contraseña', 'type' => 'password' ],
             ],
         ]);
     }
@@ -84,8 +101,15 @@ class UserController extends Controller
     }
 
 
-    public function destroy(User $user)
+    public function destroy(int $id)
     {
+        $user = User::withTrashed()->find($id);
+
+        if ($user->trashed()) {
+            $user->restore();
+            return response()->json(['message' => 'User restored'], 200);
+        }
+
         $user->delete();
         return response()->json(['message' => 'User deleted'], 200);
     }
